refactor: refresh token 관련 type 지정
