+++ frontend/src/hooks/useApiErrorHandler.tsx
@@ -2,8 +2,8 @@ import { AxiosError } from "axios";
 import { useNavigate } from "react-router-dom";
 
 import { useSnackbar } from "@/context/SnackbarContext";
-import useCreateRenewalToken from "@/hooks/useCreateRenewalToken";
 import useRetryMutate from "@/hooks/useRetryMutate";
+import useCreateRenewalToken from "@/hooks/api/member/useCreateRenewalToken";
 
 import { getCookie } from "@/util/cookie";
 import { COOKIE_KEY } from "@/constants";
