+++ frontend/src/pages/RollingpaperPage/components/Like.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import styled from "@emotion/styled";
 
 import useValidateParam from "@/hooks/useValidateParam";
